# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Digital Products Telegram Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  deliveryContent: string;
  isActive: boolean;
  stock: number;
  createdAt: string;
}

export interface CreateProductBody {
  name: string;
  description: string;
  price: number;
  deliveryContent: string;
  isActive?: boolean;
  stock?: number;
}

export interface UpdateProductBody {
  name?: string;
  description?: string;
  price?: number;
  deliveryContent?: string;
  isActive?: boolean;
  stock?: number;
}

export interface Order {
  id: number;
  telegramUserId: string;
  /** @nullable */
  telegramUsername: string | null;
  /** @nullable */
  telegramFirstName: string | null;
  productId: number;
  productName: string;
  productPrice: number;
  status: string;
  paymentMethod: string;
  /** @nullable */
  paymentProofUrl: string | null;
  /** @nullable */
  deliveryMessage: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ConfirmOrderBody {
  deliveryMessage: string;
}

export interface Settings {
  id: number;
  /** @nullable */
  qrImageUrl: string | null;
  paymentInstructions: string;
  welcomeMessage: string;
  isOpen: boolean;
  closedMessage: string;
}

export interface UpdateSettingsBody {
  qrImageUrl?: string;
  paymentInstructions?: string;
  welcomeMessage?: string;
  isOpen?: boolean;
  closedMessage?: string;
}

export interface BroadcastBody {
  message: string;
}

export type ListOrdersParams = {
  status?: string;
};