# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Digital Products Telegram Bot API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all products
 */
export const ListProductsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  price: zod.number(),
  deliveryContent: zod.string(),
  isActive: zod.boolean(),
  createdAt: zod.date(),
});
export const ListProductsResponse = zod.array(ListProductsResponseItem);

/**
 * @summary Create a product
 */
export const CreateProductBody = zod.object({
  name: zod.string(),
  description: zod.string(),
  price: zod.number(),
  costPrice: zod.number().optional().default(0),
  deliveryContent: zod.string(),
  isActive: zod.boolean().optional(),
  stock: zod.number().int().optional(),
  validPeriod: zod.string().optional().nullable(),
  warranty: zod.string().optional().nullable(),
  categoryId: zod.number().int().optional().nullable(),
  installGuide: zod.string().optional().nullable(),
});

/**
 * @summary Get a product
 */
export const GetProductParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProductResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  price: zod.number(),
  deliveryContent: zod.string(),
  isActive: zod.boolean(),
  createdAt: zod.date(),
});

/**
 * @summary Update a product
 */
export const UpdateProductParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProductBody = zod.object({
  name: zod.string().optional(),
  description: zod.string().optional(),
  price: zod.number().optional(),
  costPrice: zod.number().optional(),
  deliveryContent: zod.string().optional(),
  isActive: zod.boolean().optional(),
  stock: zod.number().int().optional(),
  validPeriod: zod.string().optional().nullable(),
  warranty: zod.string().optional().nullable(),
  categoryId: zod.number().int().optional().nullable(),
  installGuide: zod.string().optional().nullable(),
});

export const UpdateProductResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  price: zod.number(),
  deliveryContent: zod.string(),
  isActive: zod.boolean(),
  createdAt: zod.date(),
});

/**
 * @summary Delete a product
 */
export const DeleteProductParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all orders
 */
export const ListOrdersQueryParams = zod.object({
  status: zod.coerce.string().optional(),
});

export const ListOrdersResponseItem = zod.object({
  id: zod.number(),
  telegramUserId: zod.string(),
  telegramUsername: zod.string().nullable(),
  telegramFirstName: zod.string().nullable(),
  productId: zod.number(),
  productName: zod.string(),
  productPrice: zod.number(),
  status: zod.string(),
  paymentProofUrl: zod.string().nullable(),
  deliveryMessage: zod.string().nullable(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});
export const ListOrdersResponse = zod.array(ListOrdersResponseItem);

/**
 * @summary Get an order
 */
export const GetOrderParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOrderResponse = zod.object({
  id: zod.number(),
  telegramUserId: zod.string(),
  telegramUsername: zod.string().nullable(),
  telegramFirstName: zod.string().nullable(),
  productId: zod.number(),
  productName: zod.string(),
  productPrice: zod.number(),
  status: zod.string(),
  paymentProofUrl: zod.string().nullable(),
  deliveryMessage: zod.string().nullable(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Confirm payment and send product
 */
export const ConfirmOrderParams = zod.object({
  id: zod.coerce.number(),
});

export const ConfirmOrderBody = zod.object({
  deliveryMessage: zod.string(),
});

export const ConfirmOrderResponse = zod.object({
  id: zod.number(),
  telegramUserId: zod.string(),
  telegramUsername: zod.string().nullable(),
  telegramFirstName: zod.string().nullable(),
  productId: zod.number(),
  productName: zod.string(),
  productPrice: zod.number(),
  status: zod.string(),
  paymentProofUrl: zod.string().nullable(),
  deliveryMessage: zod.string().nullable(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Reject an order
 */
export const RejectOrderParams = zod.object({
  id: zod.coerce.number(),
});

export const RejectOrderResponse = zod.object({
  id: zod.number(),
  telegramUserId: zod.string(),
  telegramUsername: zod.string().nullable(),
  telegramFirstName: zod.string().nullable(),
  productId: zod.number(),
  productName: zod.string(),
  productPrice: zod.number(),
  status: zod.string(),
  paymentProofUrl: zod.string().nullable(),
  deliveryMessage: zod.string().nullable(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Get bot settings
 */
export const GetSettingsResponse = zod.object({
  id: zod.number(),
  qrImageUrl: zod.string().nullable(),
  paymentInstructions: zod.string(),
  welcomeMessage: zod.string(),
});

/**
 * @summary Update bot settings
 */
export const UpdateSettingsBody = zod.object({
  qrImageUrl: zod.string().optional(),
  paymentInstructions: zod.string().optional(),
  welcomeMessage: zod.string().optional(),
  isOpen: zod.boolean().optional(),
  closedMessage: zod.string().optional(),
  requiredChannel: zod.string().optional().nullable(),
  channelInviteLink: zod.string().optional().nullable(),
});

export const UpdateSettingsResponse = zod.object({
  id: zod.number(),
  qrImageUrl: zod.string().nullable(),
  paymentInstructions: zod.string(),
  welcomeMessage: zod.string(),
  isOpen: zod.boolean(),
  closedMessage: zod.string(),
  requiredChannel: zod.string().nullable(),
  channelInviteLink: zod.string().nullable(),
});