# lib/api-zod/src/generated/types/confirmOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Digital Products Telegram Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface ConfirmOrderBody {
  deliveryMessage: string;
}