# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Digital Products Telegram Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: number;
  telegramUserId: string;
  /** @nullable */
  telegramUsername: string | null;
  /** @nullable */
  telegramFirstName: string | null;
  productId: number;
  productName: string;
  productPrice: number;
  status: string;
  /** @nullable */
  paymentProofUrl: string | null;
  /** @nullable */
  deliveryMessage: string | null;
  createdAt: Date;
  updatedAt: Date;
}