# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Digital Products Telegram Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  costPrice?: number;
  deliveryContent: string;
  isActive: boolean;
  stock: number;
  validPeriod?: string | null;
  warranty?: string | null;
  createdAt: Date;
}