# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Digital Products Telegram Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  id: number;
  /** @nullable */
  qrImageUrl: string | null;
  paymentInstructions: string;
  welcomeMessage: string;
}